modificate files producto services

+++ api-rest/inventario/internal/entitys/product.go
@@ -4,6 +4,6 @@ type Producto struct {
 	ID          int64   `db:"id" json:"id"`
 	Name        string  `db:"name" json:"name"`
 	Description string  `db:"description" json:"description"`
-	Price       float64 `db:"price" json:"price"`
+	Price       float32 `db:"price" json:"price"`
 	Created_by  int64   `db:"created_by" json:"created_by"`
 }
